requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
stripe
itsdangerous
werkzeug
email-validator
pytest
httpx
